+++ glue/core/tests/test_data_combo_helper.py
@@ -324,3 +324,22 @@ def test_component_id_combo_helper_reorder():
     data.reorder_components(data.components[::-1])
 
     assert selection_choices(state, 'combo') == "y:x"
+
+
+def test_component_id_combo_helper_none():
+
+    # Make sure the none=True option works
+
+    state = ExampleState()
+
+    data = Data(x=[1, 2, 3], y=[2, 3, 4], label='data1')
+    dc = DataCollection([data])
+
+    helper = ComponentIDComboHelper(state, 'combo', dc)  # noqa
+    helper.append_data(data)
+
+    assert selection_choices(state, 'combo') == "x:y"
+
+    helper.none = True
+
+    assert selection_choices(state, 'combo') == ":x:y"